web-app/requirements.txt
flask
pymongo
python-dotenv
Pillow
requests
